+++ package.json
@@ -8,7 +8,9 @@
   },
   "dependencies": {
     "react": "16.6.3",
-    "react-native": "0.57.8"
+    "react-native": "0.57.8",
+    "react-native-gesture-handler": "^1.0.12",
+    "react-navigation": "^3.0.8"
   },
   "devDependencies": {
     "babel-jest": "23.6.0",
